매니저 수정페이지 작업
