+++ .gitignore
@@ -78,14 +78,4 @@ frontend/*.sw?
 
 application-secrets.properties
 office24-user.sql
-
-# ============================================================================
-
 backend/img
-
-# ============================================================================
-
-temp.sql
-.env
-
-# ============================================================================
