+++ backend/src/main/java/com/kh/backend/booking/BookingMapper.java
@@ -8,8 +8,6 @@ import java.util.Map;
 
 @Mapper
 public interface BookingMapper {
-    
     List<Map<String, Object>> selectBookingsByManager(@Param("no") Integer no, @Param("limit") int limit, @Param("offset") int offset);
-
     int countBookingsByManager(@Param("no") Integer no);
 }
