+++ backend/src/main/java/com/kh/backend/manager/ManagerMapper.java
@@ -4,8 +4,6 @@ import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
 public interface ManagerMapper {
-    // 등록
     void insertManager(Manager manager);
-    // 조회
     Manager findById(String id);
 }
